+++ examples/style-base/src/index.vue
@@ -1,4 +1,13 @@
-<style>
+<template>
+  <div>
+    asdasdasd
+  </div>
+</template>
+
+<script>
+</script>
+
+<style lang="less">
 .one {
   display: block;
   width: 100px;