+++ examples/vue-v3/src/test.vue
@@ -1,14 +1,12 @@
+<script setup>
+const message = ref('a')
+</script>
+
 <template>
   <div>
     asdasdasd
   </div>
 </template>
-
-<script setup>
-
-const message = ref('a')
-
-</script>
 
 <style lang="scss" scoped>