+++ packages/eslint-config-base/index.js
@@ -4,8 +4,10 @@ module.exports = {
     browser: true,
     node: true,
   },
+  reportUnusedDisableDirectives: true,
   extends: [
-    "standard",
+    "./extend-standard",
+    "plugin:import/recommended",
     "plugin:eslint-comments/recommended",
     "plugin:jsonc/recommended-with-jsonc",
     "plugin:yml/standard",
@@ -16,6 +18,7 @@ module.exports = {
       node: { extensions: [".js", ".mjs", ".ts", ".d.ts"] },
     },
   },
+  plugins: ["html", "n", "unicorn"],
   rules: {
     "eslint-comments/disable-enable-pair": "off",
     // Common
@@ -118,6 +121,26 @@ module.exports = {
     {
       files: ["*.json", "*.json5"],
       parser: "jsonc-eslint-parser",
+      rules: {
+        "jsonc/array-bracket-spacing": ["error", "never"],
+        "jsonc/comma-dangle": ["error", "never"],
+        "jsonc/comma-style": ["error", "last"],
+        "jsonc/indent": ["error", 2],
+        "jsonc/key-spacing": [
+          "error",
+          { beforeColon: false, afterColon: true },
+        ],
+        "jsonc/no-octal-escape": "error",
+        "jsonc/object-curly-newline": [
+          "error",
+          { multiline: true, consistent: true },
+        ],
+        "jsonc/object-curly-spacing": ["error", "always"],
+        "jsonc/object-property-newline": [
+          "error",
+          { allowMultiplePropertiesPerLine: true },
+        ],
+      },
     },
     {
       files: ["package.json"],
@@ -128,6 +151,7 @@ module.exports = {
           {
             pathPattern: "^$",
             order: [
+              "publisher",
               "name",
               "version",
               "description",
@@ -157,6 +181,10 @@ module.exports = {
             pathPattern: "^(?:dev|peer|optional|bundled)?[Dd]ependencies$",
             order: { type: "asc" },
           },
+          {
+            pathPattern: "^exports.*$",
+            order: ["types", "require", "import"],
+          },
         ],
       },
     },
@@ -180,6 +208,24 @@ module.exports = {
       files: ["*.test.ts", "*.test.js", "*.spec.ts", "*.spec.js"],
       rules: {
         "no-unused-expressions": "off",
+      },
+    },
+    {
+      // Code blocks in markdown file
+      files: ["**/*.md/*.*"],
+      rules: {
+        "@typescript-eslint/no-redeclare": "off",
+        "@typescript-eslint/no-unused-vars": "off",
+        "@typescript-eslint/no-use-before-define": "off",
+        "@typescript-eslint/no-var-requires": "off",
+        "@typescript-eslint/comma-dangle": "off",
+        "import/no-unresolved": "off",
+        "no-alert": "off",
+        "no-console": "off",
+        "no-restricted-imports": "off",
+        "no-undef": "off",
+        "no-unused-expressions": "off",
+        "no-unused-vars": "off",
       },
     },
   ],