+++ packages/eslint-config-base/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @realive/eslint-config-base
 
+## 2.0.0
+
+### Major Changes
+
+- 更新配置
+
 ## 1.1.0
+
 ### Minor Changes
 
 - 发布 lint