+++ packages/eslint-config-ts/CHANGELOG.md
@@ -1,6 +1,18 @@
 # @realive/eslint-config-ts
 
+## 2.0.0
+
+### Major Changes
+
+- 更新配置
+
+### Patch Changes
+
+- Updated dependencies
+  - @realive/eslint-config-base@2.0.0
+
 ## 1.1.0
+
 ### Minor Changes
 
 - 发布 lint