+++ packages/eslint-config-vue-v2/CHANGELOG.md
@@ -1,6 +1,14 @@
 # @realive/eslint-config-vue-v2
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @realive/eslint-config-ts@2.0.0
+
 ## 1.1.0
+
 ### Minor Changes
 
 - 发布 lint