+++ packages/stylelint-config-vue/CHANGELOG.md
@@ -1,6 +1,24 @@
 # @realive/stylelint-config-vue
 
+## 2.0.0
+
+### Major Changes
+
+- 更新配置
+
+### Patch Changes
+
+- Updated dependencies
+  - @realive/stylelint-config-base@2.0.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 支持 less、scss、sass、stylus、sugarss
+
 ## 1.1.0
+
 ### Minor Changes
 
 - 发布 lint